+++ CHANGELOG.rst
@@ -6,7 +6,10 @@ v0.9.0 - unreleased
 
 New Features
 ^^^^^^^^^^^^
-
+- Create :py:class:`mesmer.core.linear_regression.LinearRegression` which encapsulates
+  ``fit``, ``predict``, etc. methods around linear regression
+  (`#134 <https://github.com/MESMER-group/mesmer/pull/134>`_).
+  By `Mathias Hauser <https://github.com/mathause>`_.
 - Add ``mesmer.core.linear_regression``: xarray wrapper for ``mesmer.core._linear_regression``.
   (`#123 <https://github.com/MESMER-group/mesmer/pull/123>`_).
   By `Mathias Hauser <https://github.com/mathause>`_.